% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference.R
\name{load_reference}
\alias{load_reference}
\title{Download and load HPA RNA-seq data}
\usage{
load_reference(
  file,
  type = "hpa_brain",
  dest_dir = tempdir(),
  overwrite = FALSE
)
}
\arguments{
\item{file}{Name of the file to be saved locally (e.g., "rna_ic.zip").}

\item{type}{Type of dataset to download. Options are:
\itemize{
\item \code{"hpa_brain"} – RNA across 13 brain regions
\item \code{"rna_sn_clusters"} – RNA across 34 single nuclei cluster types in 11 brain regions
\item \code{"rna_ic"} – RNA expression in immune cell types
\item \code{"rna_ic_monaco"} - RNA expression in subsets of immune cell types
}}

\item{dest_dir}{Local directory to store the downloaded file (default: \code{tempdir()}).}

\item{overwrite}{Logical; if \code{TRUE}, the file will be re-downloaded even if it exists.}
}
\value{
A \code{data.frame} containing the unzipped RNA-seq dataset.
}
\description{
Downloads and reads a compressed RNA-seq dataset from the Human Protein Atlas.
}
